pluto-tools/requirements.txt
numpy
pyadi-iio
pyrtlsdr
scipy
